Add `Target` Constructor (#3815)
