+++ examples/sync/src/databases/keyless.rs
@@ -164,10 +164,7 @@ where
     E: Storage + Clock + Metrics,
 {
     async fn current_target(&self) -> compact::Target<Self::Family, Key> {
-        compact::Target {
-            root: self.root(),
-            leaf_count: self.bounds().await.end,
-        }
+        compact::Target::new(self.root(), self.bounds().await.end)
     }
 }
