+++ storage/fuzz/fuzz_targets/qmdb_any_fixed_sync.rs
@@ -11,6 +11,7 @@ use commonware_storage::{
     merkle::{full::Config as MerkleConfig, mmb, mmr, Family as MerkleFamily},
     qmdb::{
         any::{
+            sync::Target,
             unordered::fixed::{Db, Operation as FixedOperation},
             FixedConfig as Config,
         },
@@ -117,7 +118,7 @@ fn test_config(test_name: &str, pooler: &impl BufferPooler) -> Config<TwoCap, Se
 async fn test_sync<F, R>(
     context: deterministic::Context,
     resolver: R,
-    target: sync::Target<F, commonware_cryptography::sha256::Digest>,
+    target: Target<F, commonware_cryptography::sha256::Digest>,
     fetch_batch_size: u64,
     test_name: &str,
     sync_id: usize,
@@ -133,7 +134,11 @@ where
     let db_config = test_config(test_name, &context);
     let expected_root = target.root;
 
-    let sync_config: sync::engine::Config<FixedDb<F>, R> = sync::engine::Config {
+    let sync_config: sync::engine::Config<
+        FixedDb<F>,
+        R,
+        Target<F, commonware_cryptography::sha256::Digest>,
+    > = sync::engine::Config {
         context: context.child("sync").with_attribute("id", sync_id),
         update_rx: None,
         finish_rx: None,
@@ -238,11 +243,10 @@ fn fuzz_family<F: MerkleFamily>(input: &mut FuzzInput, test_name: &str) {
                         .await
                         .expect("commit should not fail");
                     db.commit().await.expect("Commit should not fail");
-                    let target = sync::Target {
-                        root: db.root(),
-                        ops_root: db.root(),
-                        range: non_empty_range!(db.sync_boundary(), db.bounds().await.end),
-                    };
+                    let target = Target::new(
+                        db.root(),
+                        non_empty_range!(db.sync_boundary(), db.bounds().await.end),
+                    );
 
                     let wrapped_src = Arc::new(db);
                     let _result = test_sync(
