+++ storage/src/qmdb/immutable/sync/tests.rs
@@ -10,12 +10,13 @@ use crate::{
     merkle::{self, full::Config as MerkleConfig, Location},
     qmdb::{
         self,
+        any::sync::Target,
         immutable::{self, variable::Operation},
         sync::{
             self,
             engine::{Config, NextStep},
             resolver::Resolver,
-            Engine, Target,
+            Engine,
         },
     },
     translator::TwoCap,
@@ -126,11 +127,10 @@ where
         let config = Config {
             db_config: db_config.clone(),
             fetch_batch_size,
-            target: Target {
-                root: target_root,
-                ops_root: target_root,
-                range: non_empty_range!(target_oldest_retained_loc, target_op_count),
-            },
+            target: Target::new(
+                target_root,
+                non_empty_range!(target_oldest_retained_loc, target_op_count),
+            ),
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -189,11 +189,10 @@ where
         let config = Config {
             db_config,
             fetch_batch_size: NZU64!(10),
-            target: Target {
-                root: target_root,
-                ops_root: target_root,
-                range: non_empty_range!(target_oldest_retained_loc, target_op_count),
-            },
+            target: Target::new(
+                target_root,
+                non_empty_range!(target_oldest_retained_loc, target_op_count),
+            ),
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -241,11 +240,7 @@ where
         let config = Config {
             db_config: db_config.clone(),
             fetch_batch_size: NZU64!(5),
-            target: Target {
-                root: target_root,
-                ops_root: target_root,
-                range: non_empty_range!(lower_bound, op_count),
-            },
+            target: Target::new(target_root, non_empty_range!(lower_bound, op_count)),
             context: client_context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -315,11 +310,10 @@ where
             let config = Config {
                 context: context.child("client"),
                 db_config: H::config(&format!("update_test_{}", context.next_u64()), &context),
-                target: Target {
-                    root: initial_root,
-                    ops_root: initial_root,
-                    range: non_empty_range!(initial_lower_bound, initial_upper_bound),
-                },
+                target: Target::new(
+                    initial_root,
+                    non_empty_range!(initial_lower_bound, initial_upper_bound),
+                ),
                 resolver: target_db.clone(),
                 fetch_batch_size: NZU64!(2),
                 max_outstanding_requests: 10,
@@ -329,7 +323,8 @@ where
                 reached_target_tx: None,
                 max_retained_roots: 1,
             };
-            let mut client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
+            let mut client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
+                Engine::new(config).await.unwrap();
             loop {
                 client = match client.step().await.unwrap() {
                     NextStep::Continue(new_client) => new_client,
@@ -343,11 +338,10 @@ where
         };
 
         update_sender
-            .send(Target {
-                root: final_root,
-                ops_root: final_root,
-                range: non_empty_range!(initial_lower_bound, final_upper_bound),
-            })
+            .send(Target::new(
+                final_root,
+                non_empty_range!(initial_lower_bound, final_upper_bound),
+            ))
             .await
             .unwrap();
 
@@ -399,11 +393,7 @@ where
         let config = Config {
             db_config: H::config(&format!("subset_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(10),
-            target: Target {
-                root: target_root,
-                ops_root: target_root,
-                range: non_empty_range!(lower_bound, op_count),
-            },
+            target: Target::new(target_root, non_empty_range!(lower_bound, op_count)),
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -455,11 +445,7 @@ where
         let config = Config {
             db_config: sync_db_config,
             fetch_batch_size: NZU64!(10),
-            target: Target {
-                root,
-                ops_root: root,
-                range: non_empty_range!(lower_bound, upper_bound),
-            },
+            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
             context: context.child("sync"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -509,11 +495,7 @@ where
         let config = Config {
             db_config: sync_config,
             fetch_batch_size: NZU64!(10),
-            target: Target {
-                root,
-                ops_root: root,
-                range: non_empty_range!(lower_bound, upper_bound),
-            },
+            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
             context: context.child("sync"),
             resolver: resolver.clone(),
             apply_batch_size: 1024,
@@ -559,11 +541,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("lb-dec-{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(5),
-            target: Target {
-                root: initial_root,
-                ops_root: initial_root,
-                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
-            },
+            target: Target::new(
+                initial_root,
+                non_empty_range!(initial_lower_bound, initial_upper_bound),
+            ),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -572,17 +553,17 @@ where
             reached_target_tx: None,
             max_retained_roots: 1,
         };
-        let client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
+        let client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
+            Engine::new(config).await.unwrap();
 
         update_sender
-            .send(Target {
-                root: initial_root,
-                ops_root: initial_root,
-                range: non_empty_range!(
+            .send(Target::new(
+                initial_root,
+                non_empty_range!(
                     initial_lower_bound.checked_sub(1).unwrap(),
                     initial_upper_bound
                 ),
-            })
+            ))
             .await
             .unwrap();
 
@@ -622,11 +603,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("ub-dec-{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(5),
-            target: Target {
-                root: initial_root,
-                ops_root: initial_root,
-                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
-            },
+            target: Target::new(
+                initial_root,
+                non_empty_range!(initial_lower_bound, initial_upper_bound),
+            ),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -635,14 +615,14 @@ where
             reached_target_tx: None,
             max_retained_roots: 1,
         };
-        let client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
+        let client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
+            Engine::new(config).await.unwrap();
 
         update_sender
-            .send(Target {
-                root: initial_root,
-                ops_root: initial_root,
-                range: non_empty_range!(initial_lower_bound, initial_upper_bound - 1),
-            })
+            .send(Target::new(
+                initial_root,
+                non_empty_range!(initial_lower_bound, initial_upper_bound - 1),
+            ))
             .await
             .unwrap();
 
@@ -696,11 +676,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("bounds_inc_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(1),
-            target: Target {
-                root: initial_root,
-                ops_root: initial_root,
-                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
-            },
+            target: Target::new(
+                initial_root,
+                non_empty_range!(initial_lower_bound, initial_upper_bound),
+            ),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 1,
@@ -711,11 +690,10 @@ where
         };
 
         update_sender
-            .send(Target {
-                root: final_root,
-                ops_root: final_root,
-                range: non_empty_range!(final_lower_bound, final_upper_bound),
-            })
+            .send(Target::new(
+                final_root,
+                non_empty_range!(final_lower_bound, final_upper_bound),
+            ))
             .await
             .unwrap();
 
@@ -769,11 +747,7 @@ where
         let config = Config {
             db_config,
             fetch_batch_size: NZU64!(100),
-            target: Target {
-                root: target_root,
-                ops_root: target_root,
-                range: non_empty_range!(bounds.start, bounds.end),
-            },
+            target: Target::new(target_root, non_empty_range!(bounds.start, bounds.end)),
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -834,11 +808,7 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("done_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(20),
-            target: Target {
-                root,
-                ops_root: root,
-                range: non_empty_range!(lower_bound, upper_bound),
-            },
+            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -851,11 +821,10 @@ where
         let synced_db: DbOf<H> = sync::sync(config).await.unwrap();
 
         let _ = update_sender
-            .send(Target {
-                root: sha256::Digest::from([2u8; 32]),
-                ops_root: sha256::Digest::from([2u8; 32]),
-                range: non_empty_range!(lower_bound + 1, upper_bound + 1),
-            })
+            .send(Target::new(
+                sha256::Digest::from([2u8; 32]),
+                non_empty_range!(lower_bound + 1, upper_bound + 1),
+            ))
             .await;
 
         assert_eq!(H::db_root(&synced_db), root);
@@ -1253,10 +1222,8 @@ mod compact_variable_mmr {
         deterministic::Runner::default().start(|_context| async move {
             let resolver: Arc<commonware_utils::sync::AsyncRwLock<Option<SourceDb>>> =
                 Arc::new(commonware_utils::sync::AsyncRwLock::new(None));
-            let target = sync::compact::Target {
-                root: sha256::Digest::from([0; 32]),
-                leaf_count: Location::new(1),
-            };
+            let target =
+                sync::compact::Target::new(sha256::Digest::from([0; 32]), Location::new(1));
 
             assert!(matches!(
                 sync::compact::Resolver::get_compact_state(&resolver, target).await,
@@ -1286,10 +1253,7 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let client_cfg = client_config(&suffix);
             let client: ClientDb = sync::compact::sync(sync::compact::Config {
@@ -1335,10 +1299,7 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1381,10 +1342,7 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1438,10 +1396,7 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1482,10 +1437,7 @@ mod compact_variable_mmr {
                 .merkleize(&source, Some(vec![1]), Location::new(1));
             source.apply_batch(batch1).await.unwrap();
             source.commit().await.unwrap();
-            let stale_target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: source.bounds().await.end,
-            };
+            let stale_target = sync::compact::Target::new(source.root(), source.bounds().await.end);
 
             let batch2 = source
                 .new_batch()
@@ -1493,10 +1445,8 @@ mod compact_variable_mmr {
                 .merkleize(&source, Some(vec![2]), Location::new(2));
             source.apply_batch(batch2).await.unwrap();
             source.commit().await.unwrap();
-            let current_target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: source.bounds().await.end,
-            };
+            let current_target =
+                sync::compact::Target::new(source.root(), source.bounds().await.end);
             assert_ne!(stale_target, current_target);
 
             let source = Arc::new(source);
@@ -1729,10 +1679,8 @@ mod compact_variable_mmb {
         deterministic::Runner::default().start(|_context| async move {
             let resolver: Arc<commonware_utils::sync::AsyncRwLock<Option<SourceDb>>> =
                 Arc::new(commonware_utils::sync::AsyncRwLock::new(None));
-            let target = sync::compact::Target {
-                root: sha256::Digest::from([0; 32]),
-                leaf_count: Location::new(1),
-            };
+            let target =
+                sync::compact::Target::new(sha256::Digest::from([0; 32]), Location::new(1));
 
             assert!(matches!(
                 sync::compact::Resolver::get_compact_state(&resolver, target).await,
@@ -1762,10 +1710,7 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let client_cfg = client_config(&suffix);
             let client: ClientDb = sync::compact::sync(sync::compact::Config {
@@ -1811,10 +1756,7 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1857,10 +1799,7 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1914,10 +1853,7 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: bounds.end,
-            };
+            let target = sync::compact::Target::new(source.root(), bounds.end);
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1958,10 +1894,7 @@ mod compact_variable_mmb {
                 .merkleize(&source, Some(vec![1]), Location::new(1));
             source.apply_batch(batch1).await.unwrap();
             source.commit().await.unwrap();
-            let stale_target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: source.bounds().await.end,
-            };
+            let stale_target = sync::compact::Target::new(source.root(), source.bounds().await.end);
 
             let batch2 = source
                 .new_batch()
@@ -1969,10 +1902,8 @@ mod compact_variable_mmb {
                 .merkleize(&source, Some(vec![2]), Location::new(2));
             source.apply_batch(batch2).await.unwrap();
             source.commit().await.unwrap();
-            let current_target = sync::compact::Target {
-                root: source.root(),
-                leaf_count: source.bounds().await.end,
-            };
+            let current_target =
+                sync::compact::Target::new(source.root(), source.bounds().await.end);
             assert_ne!(stale_target, current_target);
 
             let source = Arc::new(source);
